refactor snapshot at engine level, and fix todo in CommandExecutor
